# entity-service/openapi.yaml
openapi: 3.0.1
info:
  title: Entity Service API
  version: 1.0.0
servers:
  - url: "{server}:{port}/"
    variables:
      server:
        default: http://localhost
      port:
        default: "8080"
paths:
...
  /cases/{id}:
    get:
      summary: Get a support case by ID.
      operationId: getCase
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The case with the given ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseView'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Case not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      summary: Update a support case.
      description: |
        Updates exactly one field of the case. Provide exactly one of: `state`, `severity`,
        `workState`, `watchList`, `assigneeEmail`, `bestCaseFixEta`, `mostLikelyFixEta`,
        or `worstCaseFixEta`. `watchList`, `assigneeEmail`, `bestCaseFixEta`,
        `mostLikelyFixEta`, and `worstCaseFixEta` are supported only when the ServiceNow data
        source is active.
      operationId: patchCase
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCaseRequest'
      responses:
        "200":
          description: Case updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateCaseResponse'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Case not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /cases:
    post:
      summary: Create a new support case.
      operationId: createCase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCaseRequest'
      responses:
        "201":
          description: Case created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateCaseResponse'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /cases/search:
    post:
      summary: Search support cases.
      operationId: searchCases
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchCasesRequest'
      responses:
        "200":
          description: Cases matching the filters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchCasesResponse'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /cases/{id}/comments:
    post:
      summary: Create a comment on a support case.
      operationId: createCaseComment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Case UUID.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCaseCommentRequest'
      responses:
        "201":
          description: Comment created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseComment'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /cases/{id}/tags:
    post:
      summary: Attach a free-text tag to a support case.
      operationId: addCaseTag
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Case UUID.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddCaseTagRequest'
      responses:
        "201":
          description: The attached tag.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Case not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /cases/{id}/tags/{tagId}:
    delete:
      summary: Detach a tag from a support case.
      operationId: removeCaseTag
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Case UUID.
        - name: tagId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Tag UUID.
      responses:
        "204":
          description: Tag detached.
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Case or tag not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /cases/{id}/activities/search:
    post:
      summary: Search the activity feed of a support case (comments, attachments, and optional field changes).
      operationId: searchCaseActivities
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Case UUID.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchCaseActivitiesRequest'
      responses:
        "200":
          description: Paginated activity feed for the case.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchCaseActivitiesResponse'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /tags/search:
    get:
      summary: Search existing case tags, not scoped to any single case.
      description: |
        Returns tags whose label matches the free-text query, for use in FE autocomplete when
        attaching a tag to a case. An empty/omitted `q` returns all known tags.
        Only supported for the ServiceNow data source.
      operationId: searchTags
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
          description: Free-text label query. Omit or leave empty to return all known tags.
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            maximum: 100
          description: Maximum number of tags to return.
      responses:
        "200":
          description: Tags matching the query.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchTagsResponse'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "503":
          description: Case tags are only supported for the ServiceNow data source.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /attachments:
    post:
      summary: Upload a file attachment to a support case.
      operationId: createCaseAttachment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAttachmentRequest'
      responses:
        "201":
          description: Attachment uploaded successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateAttachmentResponse'
        "400":
          description: Bad request (missing fields, invalid data URI, or file exceeds 10 MB).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Case not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /attachments/search:
    post:
      summary: Search attachments linked to a support case.
      operationId: searchCaseAttachments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchAttachmentsRequest'
      responses:
        "200":
          description: Paginated list of attachments for the case.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchAttachmentsResponse'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /attachments/{id}/content:
    get:
      summary: Download the binary content of a case attachment.
      operationId: getCaseAttachmentContent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Attachment UUID.
      responses:
        "200":
          description: Raw file content. Content-Type reflects the actual file type (e.g. image/png, application/pdf).
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Attachment not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /attachments/{id}:
    delete:
      summary: Delete a case attachment (ServiceNow data source only).
      operationId: deleteCaseAttachment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Attachment UUID.
      responses:
        "200":
          description: Attachment deleted successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteAttachmentResponse'
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Attachment not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    UpdateCaseResponse:
      type: object
      required: [message, case]
      properties:
        message:
          type: string
        case:
          $ref: '#/components/schemas/UpdatedCase'
...
    CreateCaseRequest:
      type: object
      required: [type, projectId, deploymentId, deployedProductId]
      properties:
        type:
          type: string
          enum: [case, service_request, security_report_analysis]
          description: |
            Case type.
            - `case`: standard support case; requires `subject`, `description`, `severity`, `issueType`.
            - `service_request`: catalog-based service request (ServiceNow only); requires `catalogId`, `catalogItemId`, `variables`.
            - `security_report_analysis`: security report (ServiceNow only); requires `subject`, `description`, and at least one entry in `attachments`.
        projectId:
          type: string
          format: uuid
        deploymentId:
          type: string
          format: uuid
        deployedProductId:
          type: string
          format: uuid
        subject:
          type: string
          description: Required for `case` and `security_report_analysis` types.
        description:
          type: string
          description: Required for `case` and `security_report_analysis` types.
        severity:
          type: string
          enum: [catastrophic, critical, high, medium, low]
          description: Required for `case` type.
        issueType:
          type: string
          enum: [error, partial_outage, performance_degradation, question, security_or_compliance, total_outage]
          description: Required for `case` type.
        catalogId:
          type: string
          format: uuid
          description: Required for `service_request` type. UUID of the catalog.
        catalogItemId:
          type: string
          format: uuid
          description: Required for `service_request` type. UUID of the catalog item.
        variables:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/Variable'
          description: Required for `service_request` type.
        attachments:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/CaseAttachment'
          description: Required for `security_report_analysis` type (at least one entry).
        relatedCaseId:
          type: string
          format: uuid
          description: Optional UUID of a related case.
        conversationId:
          type: string
          format: uuid
          description: Optional UUID of a related conversation.
        watchList:
          type: array
          items:
            type: string
          description: Optional list of user IDs to notify.
    CreateCaseResponse:
      type: object
      properties:
        message:
          type: string
        case:
          type: object
          properties:
            id:
              type: string
            internalId:
              type: string
            number:
              type: string
            createdBy:
              type: string
            createdOn:
              type: string
              format: date-time
            state:
              type: string
...
    CaseView:
      type: object
      properties:
        id:
          type: string
          format: uuid
        number:
          type: string
        internalId:
          type: string
        subject:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [catastrophic, critical, high, medium, low]
        issueType:
          type: string
          enum: [error, partial_outage, performance_degradation, question, security_or_compliance, total_outage]
        state:
          type: string
          enum: [open, work_in_progress, waiting_on_wso2, awaiting_info, reopened, solution_proposed, closed]
        workState:
          type: string
          enum: [ongoing, paused]
          nullable: true
          description: Work sub-state, present only when state is work_in_progress.
        type:
          type: string
          nullable: true
          description: Case type label. Populated for ServiceNow cases; null for Postgres cases.
        engagementType:
          type: string
          nullable: true
          description: Engagement type label. Populated for ServiceNow cases; null for Postgres cases.
        createdOn:
          type: string
          format: date-time
        updatedOn:
          type: string
          format: date-time
        closedOn:
          type: string
          format: date-time
          nullable: true
          description: Timestamp when the case was closed. null if not yet closed.
        createdBy:
          $ref: '#/components/schemas/UserRef'
        createdByUser:
          nullable: true
          $ref: '#/components/schemas/UserReference'
        project:
          $ref: '#/components/schemas/EntityRef'
        deployment:
          nullable: true
          $ref: '#/components/schemas/EntityRef'
          description: Deployment associated with the case. null for ServiceNow cases with no deployment.
        deployedProduct:
          $ref: '#/components/schemas/DeployedProductRef'
        product:
          $ref: '#/components/schemas/EntityRef'
        catalog:
          nullable: true
          $ref: '#/components/schemas/EntityRef'
          description: Service catalog entry. Populated for ServiceNow cases; null for Postgres cases.
        catalogItem:
          nullable: true
          $ref: '#/components/schemas/EntityRef'
          description: Service catalog item. Populated for ServiceNow cases; null for Postgres cases.
        assignedTeam:
          nullable: true
          $ref: '#/components/schemas/EntityRef'
          description: Team assigned to the case. Populated for ServiceNow cases; null for Postgres cases.
        conversation:
          nullable: true
          $ref: '#/components/schemas/EntityRef'
          description: Linked conversation. Populated for ServiceNow cases; null for Postgres cases.
        assignedEngineer:
          nullable: true
          $ref: '#/components/schemas/AssignedEngineerRef'
        assignedEngineerUser:
          nullable: true
          $ref: '#/components/schemas/UserReference'
        parentCase:
          nullable: true
          $ref: '#/components/schemas/CaseNumberRef'
        relatedCase:
          nullable: true
          $ref: '#/components/schemas/CaseNumberRef'
        account:
          nullable: true
          $ref: '#/components/schemas/AccountRef'
        linkedChangeRequests:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/LinkedChangeRequestRef'
          description: >-
            Change requests raised from this case, in no particular order. Only
            service-request cases carry them; absent or empty for every other case
            type. This is a one-to-many relationship: promoting the same change
            through each environment produces one change request per environment,
            all pointing back at the same service request.
        resolvedOn:
          type: string
          format: date-time
          nullable: true
          description: Timestamp when the case was resolved. Populated for resolved/closed ServiceNow cases; null otherwise.
        resolutionCode:
          type: string
          nullable: true
          enum: [SOLVED_FIXED_BY_SUPPORT_GUIDANCE_PROVIDED, SOLVED_FIXED_BY_CLOSING_RELATED_INCIDENT, SOLVED_FIXED_BY_CLOSING_RELATED_RD_TICKET, SOLVED_WORKAROUND_PROVIDED, SOLVED_BY_CUSTOMER, CONSIDERED_FOR_ROADMAP, INCONCLUSIVE_OUT_OF_SCOPE, INCONCLUSIVE_CANNOT_REPRODUCE, INCONCLUSIVE_NO_WORKAROUND, DUPLICATE_ISSUE, VOIDED_CANCELED, ON_HOLD, CONSIDERED_FOR_ROADMAP_ALT, SOLVED_FIXED_THE_ISSUE, SOLVED_WORKAROUND_PROVIDED_ALT, SOLVED_BY_CONTRIBUTOR, SOLVED_BY_NOVERA, ABRUPTLY_CLOSED_DUE_TO_NON_RESPONSIVENESS]
          description: Resolution code. Populated for resolved/closed ServiceNow cases; null otherwise.
        cause:
          type: string
          nullable: true
          enum: [SOLUTION_ARCHITECTURE, DEPLOYMENT_ARCHITECTURE, USER_ERROR_CONFIGURATION, USER_ERROR_PRODUCT_CONCEPT, USER_ERROR_RUNTIME, USER_ERROR_RECOMMENDATION_BEST_PRACTICES, CUSTOMIZATION_LIMITATION, CUSTOMIZATION_BUG, DOCUMENTATION_GAP, DOCUMENTATION_ERROR, PRODUCT_LIMITATION, PRODUCT_BUG, PRODUCT_REGRESSION, PRODUCT_MIGRATION, INFRASTRUCTURE_DATABASE, INFRASTRUCTURE_OS, INFRASTRUCTURE_NETWORK, INFRASTRUCTURE_JDK, INFRASTRUCTURE_LDAP, INFRASTRUCTURE_LOAD_BALANCER, INFRASTRUCTURE_IAAS, INFRASTRUCTURE_EXTERNAL_PRODUCT, INFRASTRUCTURE_PROXY, INFRASTRUCTURE_OTHER, UNKNOWN]
          description: Root-cause category. Populated for resolved/closed ServiceNow cases; null otherwise.
        resolutionNotes:
          type: string
          nullable: true
          description: Free-text resolution/close notes. Populated for resolved/closed ServiceNow cases; null otherwise.
        bestCaseFixEta:
          type: string
          format: date
          nullable: true
          description: >-
            Internal-only best-case fix-commitment date (date-only, YYYY-MM-DD), visible to
            CSM engineers only. Populated for ServiceNow cases; null otherwise.
        mostLikelyFixEta:
          type: string
          format: date
          nullable: true
          description: >-
            Internal-only most-likely fix-commitment date (date-only, YYYY-MM-DD), visible
            to CSM engineers only. Populated for ServiceNow cases; null otherwise.
        worstCaseFixEta:
          type: string
          format: date
          nullable: true
          description: >-
            Internal-only worst-case fix-commitment date (date-only, YYYY-MM-DD), visible to
            CSM engineers only. Populated for ServiceNow cases; null otherwise.
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
...